# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH 26128 Livestock Surveillance API",
    "description": "Checkpoint 3 livestock-health reporting, human review, privacy-preserving surveillance, and verified-only model governance API. This is a decision-support prototype, not a diagnostic system or clinically validated deployment.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/sync/batch": {
      "post": {
        "tags": [
          "reports and synchronization"
        ],
        "summary": "Sync Batch",
        "operationId": "sync_batch_api_v1_sync_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reports": {
      "post": {
        "tags": [
          "reports and synchronization"
        ],
        "summary": "Create Report Online",
        "operationId": "create_report_online_api_v1_reports_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Client-Mutation-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Client-Mutation-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncMutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "reports and synchronization"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/timeline": {
      "get": {
        "tags": [
          "reports and synchronization"
        ],
        "summary": "Report Timeline",
        "operationId": "report_timeline_api_v1_reports__report_id__timeline_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimelineEvent"
                  },
                  "title": "Response Report Timeline Api V1 Reports  Report Id  Timeline Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases": {
      "get": {
        "tags": [
          "reports and synchronization"
        ],
        "summary": "Vet Cases",
        "operationId": "vet_cases_api_v1_vet_cases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Vet Cases Api V1 Vet Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "tags": [
          "reports and synchronization"
        ],
        "summary": "Audit Viewer",
        "operationId": "audit_viewer_api_v1_admin_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Audit Viewer Api V1 Admin Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "farm_id": {
            "type": "string",
            "format": "uuid",
            "title": "Farm Id"
          },
          "animal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Animal Id"
          },
          "herd_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Herd Id"
          },
          "reporter_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reporter User Id"
          },
          "species": {
            "type": "string",
            "title": "Species"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "age_band": {
            "type": "string",
            "title": "Age Band"
          },
          "symptom_onset_at": {
            "type": "string",
            "format": "date-time",
            "title": "Symptom Onset At"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "appetite": {
            "type": "string",
            "title": "Appetite"
          },
          "water_intake": {
            "type": "string",
            "title": "Water Intake"
          },
          "mobility": {
            "type": "string",
            "title": "Mobility"
          },
          "respiration": {
            "type": "string",
            "title": "Respiration"
          },
          "visible_lesions": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Lesions"
          },
          "discharge": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discharge"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "vaccination_status": {
            "type": "string",
            "title": "Vaccination Status"
          },
          "recent_movement": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Movement"
          },
          "recent_contact": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Contact"
          },
          "mortality_count": {
            "type": "integer",
            "title": "Mortality Count"
          },
          "village_name": {
            "type": "string",
            "title": "Village Name"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "location_precision": {
            "type": "string",
            "title": "Location Precision"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "voice_transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Transcript"
          },
          "optional_provider_status": {
            "additionalProperties": true,
            "type": "object",
            "title": "Optional Provider Status"
          },
          "consent_given": {
            "type": "boolean",
            "title": "Consent Given"
          },
          "consent_version": {
            "type": "string",
            "title": "Consent Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "sync_status": {
            "type": "string",
            "title": "Sync Status"
          },
          "client_mutation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Mutation Id"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "created_at_device": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Device"
          },
          "received_at_server": {
            "type": "string",
            "format": "date-time",
            "title": "Received At Server"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "farm_id",
          "animal_id",
          "herd_id",
          "reporter_user_id",
          "species",
          "language",
          "age_band",
          "symptom_onset_at",
          "severity",
          "appetite",
          "water_intake",
          "mobility",
          "respiration",
          "visible_lesions",
          "discharge",
          "temperature_c",
          "vaccination_status",
          "recent_movement",
          "recent_contact",
          "mortality_count",
          "village_name",
          "latitude",
          "longitude",
          "location_precision",
          "notes",
          "voice_transcript",
          "optional_provider_status",
          "consent_given",
          "consent_version",
          "status",
          "sync_status",
          "client_mutation_id",
          "idempotency_key",
          "created_at_device",
          "received_at_server",
          "updated_at",
          "version"
        ],
        "title": "ReportOut"
      },
      "ReportPayload": {
        "properties": {
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "location_precision": {
            "type": "string",
            "enum": [
              "EXACT",
              "APPROXIMATE",
              "VILLAGE_ONLY"
            ],
            "title": "Location Precision",
            "default": "VILLAGE_ONLY"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "farm_id": {
            "type": "string",
            "format": "uuid",
            "title": "Farm Id"
          },
          "animal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Animal Id"
          },
          "herd_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Herd Id"
          },
          "species": {
            "type": "string",
            "enum": [
              "CATTLE",
              "BUFFALO"
            ],
            "title": "Species"
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "mr",
              "hi"
            ],
            "title": "Language",
            "default": "en"
          },
          "age_band": {
            "type": "string",
            "enum": [
              "CALF",
              "YOUNG",
              "ADULT",
              "UNKNOWN"
            ],
            "title": "Age Band"
          },
          "symptom_onset_at": {
            "type": "string",
            "format": "date-time",
            "title": "Symptom Onset At"
          },
          "severity": {
            "type": "string",
            "enum": [
              "MILD",
              "MODERATE",
              "SEVERE"
            ],
            "title": "Severity"
          },
          "appetite": {
            "type": "string",
            "enum": [
              "NORMAL",
              "REDUCED",
              "NONE",
              "UNKNOWN"
            ],
            "title": "Appetite"
          },
          "water_intake": {
            "type": "string",
            "enum": [
              "NORMAL",
              "REDUCED",
              "NONE",
              "UNKNOWN"
            ],
            "title": "Water Intake"
          },
          "mobility": {
            "type": "string",
            "enum": [
              "NORMAL",
              "LIMPING",
              "UNABLE_TO_STAND",
              "UNKNOWN"
            ],
            "title": "Mobility"
          },
          "respiration": {
            "type": "string",
            "enum": [
              "NORMAL",
              "DIFFICULT",
              "RAPID",
              "UNKNOWN"
            ],
            "title": "Respiration"
          },
          "visible_lesions": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Lesions"
          },
          "discharge": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discharge"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": 25.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "vaccination_status": {
            "type": "string",
            "enum": [
              "CURRENT",
              "OVERDUE",
              "UNKNOWN"
            ],
            "title": "Vaccination Status"
          },
          "recent_movement": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Movement"
          },
          "recent_contact": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Contact"
          },
          "mortality_count": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Mortality Count",
            "default": 0
          },
          "village_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Village Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "media_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Media Refs"
          },
          "voice_transcript": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Transcript"
          },
          "consent_given": {
            "type": "boolean",
            "const": true,
            "title": "Consent Given"
          },
          "consent_version": {
            "type": "string",
            "maxLength": 32,
            "pattern": "^CP1-\\d+$",
            "title": "Consent Version"
          },
          "created_at_device": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Device"
          },
          "optional_provider_status": {
            "$ref": "#/components/schemas/OptionalProviderStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "farm_id",
          "species",
          "age_band",
          "symptom_onset_at",
          "severity",
          "appetite",
          "water_intake",
          "mobility",
          "respiration",
          "vaccination_status",
          "village_name",
          "consent_given",
          "consent_version",
          "created_at_device"
        ],
        "title": "ReportPayload"
      },
...
      "SyncBatchRequest": {
        "properties": {
          "mutations": {
            "items": {
              "$ref": "#/components/schemas/SyncMutationIn"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Mutations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mutations"
        ],
        "title": "SyncBatchRequest"
      },
      "SyncBatchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SyncMutationResult"
            },
            "type": "array",
            "title": "Results"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "next_retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Retry After Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "results",
          "server_time"
        ],
        "title": "SyncBatchResponse"
      },
...
      "SyncMutationResult": {
        "properties": {
          "client_mutation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Mutation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPLIED",
              "DUPLICATE",
              "CONFLICT",
              "REJECTED"
            ],
            "title": "Status"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Version"
          },
          "received_at_server": {
            "type": "string",
            "format": "date-time",
            "title": "Received At Server"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MutationError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_mutation_id",
          "status",
          "resource_id",
          "resource_version",
          "received_at_server"
        ],
        "title": "SyncMutationResult"
      },
      "TimelineEvent": {
        "properties": {
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "occurred_at",
          "event_type",
          "actor_user_id",
          "details"
        ],
        "title": "TimelineEvent"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}